refactor!: nest HTML DOM helpers under Html2rss::Html
