+++ lib/html2rss/auto_source/scraper/discovery.rb
@@ -6,7 +6,7 @@ module Html2rss
       ##
       # Card/list discovery helpers used by AutoSource scrapers before field extraction.
       #
-      # HtmlExtractor fills article fields from a container; discovery finds those containers.
+      # Html2rss::Html::Extractor fills article fields from a container; discovery finds those containers.
       module Discovery
       end
     end
