+++ lib/html2rss/html.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Html2rss
+  ##
+  # DOM helpers for navigating and extracting from HTML trees.
+  module Html
+  end
+end
