+++ lib/html2rss/html/extractor/date_extractor.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+module Html2rss
+  module Html
+    class Extractor
+      # Extracts the earliest date from an article_tag.
+      class DateExtractor
+        # @param article_tag [Nokogiri::XML::Element] article container node
+        # @return [DateTime, nil]
+        def self.call(article_tag)
+          times = article_tag.css('[datetime]').filter_map do |tag|
+            DateTime.parse(tag['datetime'])
+          rescue ArgumentError, TypeError
+            nil
+          end
+
+          times.min
+        end
+      end
+    end
+  end
+end
