--- lib/html2rss/html_navigator.rb
@@ -1,138 +1,0 @@
-# frozen_string_literal: true
-
-module Html2rss
-  ##
-  # HtmlNavigator owns DOM chrome recognition and node traversal helpers.
-  class HtmlNavigator
-    # Heading tags used to prioritize title extraction and container assessment.
-    HEADING_TAGS = %w[h1 h2 h3 h4 h5 h6].freeze
-
-    # Element tags that indicate ignored DOM chrome when found in a container path.
-    IGNORED_CONTAINER_TAGS = %w[nav footer header svg script style].to_set.freeze
-
-    # Anchor selector used to identify the canonical article link element.
-    MAIN_ANCHOR_SELECTOR = begin
-      buf = +'a[href]:not([href=""])'
-      %w[# javascript: mailto: tel: file:// sms: data:].each do |prefix|
-        buf << %[:not([href^="#{prefix}"])]
-      end
-      buf.freeze
-    end
-
-    class << self
-      ##
-      # Extracts visible text from a given node and its children.
-      #
-      # @param tag [Nokogiri::XML::Node] the node from which to extract visible text
-      # @param separator [String] separator used to join text fragments (default is a space)
-      # @param exclude_nodes [Array<Nokogiri::XML::Node>, nil] nodes to exclude from extraction
-      # @return [String, nil] the concatenated visible text, or nil if none is found
-      def extract_visible_text(tag, separator: ' ', exclude_nodes: nil)
-        TextExtractor.call(tag, separator:, exclude_nodes:)
-      end
-
-      ##
-      # @param article_tag [Nokogiri::XML::Node] article-like container to search within
-      # @return [Nokogiri::XML::Node, nil] first eligible descendant anchor
-      def main_anchor_for(article_tag)
-        return article_tag if article_tag.name == 'a' && article_tag.matches?(MAIN_ANCHOR_SELECTOR)
-
-        article_tag.at_css(MAIN_ANCHOR_SELECTOR)
-      end
-
-      ##
-      # @param node [Nokogiri::XML::Node]
-      # @param cache [Hash, nil] identity cache used to store results (must use compare_by_identity)
-      # @return [Boolean] true when the node belongs to ignored DOM chrome
-      # rubocop:disable Metrics/CyclomaticComplexity, Metrics/MethodLength, Metrics/PerceivedComplexity
-      def ignored_container_path?(node, cache = nil)
-        return cache[node] if cache&.key?(node)
-
-        curr = node
-        visited = []
-        is_ignored = false
-
-        while curr.respond_to?(:parent) && curr
-          if cache&.key?(curr)
-            is_ignored = cache[curr]
-            break
-          end
-
-          if IGNORED_CONTAINER_TAGS.include?(curr.name)
-            is_ignored = true
-            break
-          end
-
-          visited << curr
-          curr = curr.parent
-        end
-        visited.each { |n| cache[n] = is_ignored } if cache
-
-        is_ignored
-      end
-      # rubocop:enable Metrics/CyclomaticComplexity, Metrics/MethodLength, Metrics/PerceivedComplexity
-
-      ##
-      # Returns the first parent that satisfies the condition.
-      # If the condition is met, it returns the node itself.
-      #
-      # @param node [Nokogiri::XML::Node] The node to start the search from.
-      # @param condition [Proc] The condition to be met.
-      # @return [Nokogiri::XML::Node, nil] The first parent that satisfies the condition.
-      def parent_until_condition(node, condition)
-        while node && !node.document? && node.name != 'html'
-          return node if condition.call(node)
-
-          node = node.parent
-        end
-      end
-
-      ##
-      # Think of it as `css_upwards` method.
-      # It searches for the closest parent that matches the given selector.
-      #
-      # @param current_tag [Nokogiri::XML::Node, nil] starting node
-      # @param selector [String] CSS selector to search upwards for
-      # @return [Nokogiri::XML::Node, nil] first matching node in upward traversal
-      def find_closest_selector_upwards(current_tag, selector)
-        while current_tag
-          found = current_tag.at_css(selector)
-          return found if found
-
-          return nil unless current_tag.respond_to?(:parent)
-
-          current_tag = current_tag.parent
-        end
-      end
-
-      ##
-      # Searches for the closest parent that matches the given tag name.
-      #
-      # @param current_tag [Nokogiri::XML::Node] starting node
-      # @param tag_name [String] tag name to find in ancestors
-      # @return [Nokogiri::XML::Node, nil] matching ancestor node
-      def find_tag_in_ancestors(current_tag, tag_name)
-        return current_tag if current_tag.name == tag_name
-
-        current_tag.ancestors(tag_name).first
-      end
-
-      ##
-      # Returns true if child_node is a descendant of parent_node.
-      # Walks up using parent pointers to avoid NodeSet allocations.
-      #
-      # @param child_node [Nokogiri::XML::Node] potential descendant
-      # @param parent_node [Nokogiri::XML::Node] potential ancestor
-      # @return [Boolean] true when child_node is a descendant of parent_node
-      def descendant_of?(child_node, parent_node)
-        curr = child_node.respond_to?(:parent) ? child_node.parent : nil
-        while curr
-          return true if curr == parent_node
-
-          curr = curr.respond_to?(:parent) ? curr.parent : nil
-        end
-        false
-      end
-    end
-  end
-end
