+++ spec/lib/html2rss/html/extractor/image_extractor_spec.rb
@@ -2,7 +2,7 @@
 
 require 'nokogiri'
 
-RSpec.describe Html2rss::HtmlExtractor::ImageExtractor do
+RSpec.describe Html2rss::Html::Extractor::ImageExtractor do
   let(:article_tag) { Nokogiri::HTML.fragment(html) }
 
   describe '.call' do
