+++ spec/lib/html2rss/html/extractor_spec.rb
@@ -2,10 +2,10 @@
 
 require 'nokogiri'
 
-RSpec.describe Html2rss::HtmlExtractor do
+RSpec.describe Html2rss::Html::Extractor do
   subject(:article_hash) { described_class.new(article_tag, base_url: 'https://example.com', selected_anchor:).call }
 
-  let(:selected_anchor) { Html2rss::HtmlNavigator.main_anchor_for(article_tag) }
+  let(:selected_anchor) { Html2rss::Html::Navigator.main_anchor_for(article_tag) }
 
   describe '#call with selected_anchor' do
     let(:article_tag) do
