+++ spec/lib/html2rss/selectors_spec.rb
@@ -159,7 +159,8 @@ RSpec.describe Html2rss::Selectors do
 
     context 'when extractor returns nil' do
       before do
-        allow(Html2rss::HtmlExtractor).to receive(:new).and_return(instance_double(Html2rss::HtmlExtractor, call: nil))
+        extractor = instance_double(Html2rss::Html::Extractor, call: nil)
+        allow(Html2rss::Html::Extractor).to receive(:new).and_return(extractor)
       end
 
       it 'returns article_hash' do
